+++ forge/src/main/java/tfar/bensfintasticsharks/datagen/ModBlockstateProvider.java
@@ -9,6 +9,7 @@ import net.minecraftforge.client.model.generators.ConfiguredModel;
 import net.minecraftforge.client.model.generators.ModelFile;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import tfar.bensfintasticsharks.BensFintasticSharks;
+import tfar.bensfintasticsharks.init.ModBlocks;
 
 public class ModBlockstateProvider extends BlockStateProvider {
     public ModBlockstateProvider(PackOutput gen, ExistingFileHelper exFileHelper) {
@@ -17,7 +18,12 @@ public class ModBlockstateProvider extends BlockStateProvider {
 
     @Override
     protected void registerStatesAndModels() {
-
+        simpleBlock(ModBlocks.ALGAE_BLOCK,
+                models().cross("algae_block", modLoc("block/algae_block")).renderType("cutout"));
+        simpleBlock(ModBlocks.LARGE_GREEN_ALGAE,
+                models().cross("large_green_algae", modLoc("block/large_green_algae")).renderType("cutout"));
+        simpleBlock(ModBlocks.LARGE_RED_ALGAE,
+                models().cross("large_red_algae", modLoc("block/large_red_algae")).renderType("cutout"));
     }
 
     protected void blockstateFromExistingModel(Block block) {