+++ forge/src/main/java/tfar/bensfintasticsharks/datagen/data/tags/ModBlockTagsProvider.java
@@ -6,6 +6,9 @@ import net.minecraftforge.common.data.BlockTagsProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import org.jetbrains.annotations.Nullable;
 import tfar.bensfintasticsharks.BensFintasticSharks;
+import tfar.bensfintasticsharks.init.ModBlocks;
+import tfar.bensfintasticsharks.init.ModTags;
+import net.minecraft.tags.BlockTags;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -16,5 +19,9 @@ public class ModBlockTagsProvider extends BlockTagsProvider {
 
     @Override
     protected void addTags(HolderLookup.Provider pProvider) {
+        tag(ModTags.Blocks.ALGAE).add(ModBlocks.ALGAE_BLOCK, ModBlocks.LARGE_GREEN_ALGAE,
+                ModBlocks.LARGE_RED_ALGAE);
+        tag(BlockTags.MINEABLE_WITH_HOE).add(ModBlocks.ALGAE_BLOCK, ModBlocks.LARGE_GREEN_ALGAE,
+                ModBlocks.LARGE_RED_ALGAE);
     }
 }